Dodanie funkcjonalności wyświetlania szczegółów Results
