+++ src/stylesheets/base/_colors.scss
@@ -3,4 +3,5 @@
   --black: #1c2321;
   --accent: #db504a;
   --gray: #a8a7b4;
+  --light-gray: #f7f7f8;
 }
