+++ src/stylesheets/components/_SearchResults.scss
@@ -34,37 +34,4 @@
     height: 1.6rem;
     color: var(--gray);
   }
-  .result {
-    display: flex;
-    justify-content: space-between;
-    margin-top: var(--margin-xl);
-    padding: 0 var(--margin-l);
-    height: 4rem;
-    box-shadow: var(--box-shadow);
-    border-radius: var(--border-radius);
-    background-color: var(--white);
-    .result-header {
-      display: flex;
-      flex-wrap: wrap;
-      align-items: center;
-      overflow: hidden;
-      margin: var(--margin-l) 0;
-      padding-left: 3rem;
-      line-height: var(--line-height-m);
-      span {
-        overflow: hidden;
-        text-overflow: ellipsis;
-        display: -webkit-box;
-        -webkit-line-clamp: 2;
-        -webkit-box-orient: vertical;
-      }
-    }
-    .result-button {
-      display: flex;
-      align-items: center;
-      button {
-        margin-right: var(--margin-l);
-      }
-    }
-  }
 }
